requirements.txt
absl-py==0.11.0
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
fsspec==0.8.7
future==0.18.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
grpcio==1.36.0
httplib2==0.19.0
idna==2.10
kiwisolver==1.3.1
lmdb==1.1.1
Markdown==3.3.4
matplotlib==3.3.4
multidict==5.1.0
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
packaging==20.9
Pillow==8.1.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.2.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.58.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==1.0.1
yarl==1.6.3
